Move hardinfo_clean_label() and hardinfo_clean_value() to util.c
